fix: wrap login search params usage with suspense
