+++ app/login/LoginPageClient.tsx
@@ -1,0 +1,199 @@
+"use client"
+
+import { useState, useEffect } from "react"
+import { useRouter, useSearchParams } from "next/navigation"
+import Link from "next/link"
+import { Button } from "@/components/ui/button"
+import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
+import { API_BASE_URL } from "@/lib/api-base"
+import { Eye, EyeOff, ArrowLeft, CheckCircle } from "lucide-react"
+
+export default function LoginPageClient() {
+  const router = useRouter()
+  const searchParams = useSearchParams()
+  const [showPassword, setShowPassword] = useState(false)
+  const [isLoading, setIsLoading] = useState(false)
+  const [showRegisteredMessage, setShowRegisteredMessage] = useState(false)
+  const [errors, setErrors] = useState<{ email?: string; password?: string; general?: string }>({})
+  
+  const [form, setForm] = useState({
+    email: "",
+    password: "",
+  })
+
+  useEffect(() => {
+    if (searchParams.get("registered") === "true") {
+      setShowRegisteredMessage(true)
+      setTimeout(() => setShowRegisteredMessage(false), 5000)
+    }
+  }, [searchParams])
+
+  const validateForm = (): boolean => {
+    const newErrors: { email?: string; password?: string } = {}
+
+    if (!form.email) {
+      newErrors.email = "이메일을 입력해주세요"
+    }
+
+    if (!form.password) {
+      newErrors.password = "비밀번호를 입력해주세요"
+    }
+
+    setErrors(newErrors)
+    return Object.keys(newErrors).length === 0
+  }
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault()
+    
+    if (!validateForm()) return
+
+    setIsLoading(true)
+    setErrors({})
+
+    try {
+      const response = await fetch(`${API_BASE_URL}/api/v1/auth/login`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        credentials: "include",
+        body: JSON.stringify({
+          email: form.email,
+          password: form.password,
+        }),
+      })
+
+      const data = await response.json()
+
+      if (!response.ok || data.code !== "SUCCESS") {
+        setErrors({ 
+          general: data.message || "로그인에 실패했습니다. 이메일과 비밀번호를 확인해주세요." 
+        })
+        return
+      }
+
+      if (data.data) {
+        localStorage.setItem("user", JSON.stringify(data.data))                           
+      }
+
+      router.push("/")
+    } catch (error) {
+      console.error("로그인 에러:", error)
+      setErrors({ 
+        general: "서버에 연결할 수 없습니다. 잠시 후 다시 시도해주세요." 
+      })
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  return (
+    <div className="min-h-screen bg-background">
+      <div className="flex min-h-screen flex-col items-center justify-center px-4 py-12">
+        {showRegisteredMessage && (
+          <div className="mb-4 w-full max-w-md rounded-lg border border-green-200 bg-green-50 p-4">
+            <div className="flex items-center gap-2 text-green-800">
+              <CheckCircle className="h-5 w-5" />
+              <p className="text-sm font-medium">회원가입이 완료되었습니다. 로그인해주세요.</p>
+            </div>
+          </div>
+        )}
+
+        <div className="mb-6 w-full max-w-md">
+          <Link
+            href="/"
+            className="inline-flex items-center text-sm text-muted-foreground transition-colors hover:text-primary"
+          >
+            <ArrowLeft className="mr-2 h-4 w-4" />
+            홈으로 돌아가기
+          </Link>
+        </div>
+
+        <Card className="w-full max-w-md">
+          <CardHeader className="text-center">
+            <Link href="/" className="mx-auto mb-4 flex items-center gap-2">
+              <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-primary">
+                <span className="text-xl font-bold text-primary-foreground">M</span>
+              </div>
+              <span className="text-2xl font-bold text-foreground">달려라 마라톤</span>
+            </Link>
+            <CardTitle className="text-2xl">로그인</CardTitle>
+            <CardDescription>
+              계정에 로그인하여 마라톤 대회에 참가하세요
+            </CardDescription>
+          </CardHeader>
+          <CardContent>
+            <form onSubmit={handleSubmit} className="flex flex-col gap-5">
+              {errors.general && (
+                <div className="rounded-lg border border-destructive/50 bg-destructive/10 p-3">
+                  <p className="text-sm text-destructive">{errors.general}</p>
+                </div>
+              )}
+
+              <div className="flex flex-col gap-2">
+                <Label htmlFor="email">이메일</Label>
+                <Input
+                  id="email"
+                  type="email"
+                  placeholder="example@email.com"
+                  value={form.email}
+                  onChange={(e) => setForm({ ...form, email: e.target.value })}
+                  className={errors.email ? "border-destructive" : ""}
+                />
+                {errors.email && (
+                  <p className="text-sm text-destructive">{errors.email}</p>
+                )}
+              </div>
+
+              <div className="flex flex-col gap-2">
+                <div className="flex items-center justify-between">
+                  <Label htmlFor="password">비밀번호</Label>
+                  <Link
+                    href="/forgot-password"
+                    className="text-sm text-primary hover:underline"
+                  >
+                    비밀번호 찾기
+                  </Link>
+                </div>
+                <div className="relative">
+                  <Input
+                    id="password"
+                    type={showPassword ? "text" : "password"}
+                    placeholder="비밀번호를 입력해주세요"
+                    value={form.password}
+                    onChange={(e) => setForm({ ...form, password: e.target.value })}
+                    className={errors.password ? "border-destructive pr-10" : "pr-10"}
+                  />
+                  <button
+                    type="button"
+                    onClick={() => setShowPassword(!showPassword)}
+                    className="absolute right-3 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
+                  >
+                    {showPassword ? <EyeOff className="h-4 w-4" /> : <Eye className="h-4 w-4" />}
+                  </button>
+                </div>
+                {errors.password && (
+                  <p className="text-sm text-destructive">{errors.password}</p>
+                )}
+              </div>
+
+              <Button type="submit" className="mt-2 w-full" disabled={isLoading}>
+                {isLoading ? "로그인 중..." : "로그인"}
+              </Button>
+
+              <p className="text-center text-sm text-muted-foreground">
+                계정이 없으신가요?{" "}
+                <Link href="/signup" className="font-medium text-primary hover:underline">
+                  회원가입
+                </Link>
+              </p>
+            </form>
+          </CardContent>
+        </Card>
+      </div>
+    </div>
+  )
+}
